+++ tests/integration/bot/minigame.py
@@ -3,7 +3,7 @@ import asyncio
 from discord import Embed, Member, Status, Message
 from distest.TestInterface import TestInterface
 from distest import run_dtest_bot, TestCollector
-from test_util import resetdb
+from brock_test_util import resetdb
 
 test_collector = TestCollector()
